+++ doc/source/changelog.rst
@@ -1,6 +1,14 @@
 CHANGELOG
 =========
 
+Unreleased
+----------
+
+Breaking changes
+^^^^^^^^^^^^^^^^
+
+- Removed the top-level ``convert`` command. Use ``wyrdcraeft source convert`` instead.
+
 1.0.0 (2026-03-01)
 ------------------